+++ ProjectICTLab/src/Controller/InputListenerListarEquipamentoFrame.java
@@ -22,10 +22,11 @@ public class InputListenerListarEquipamentoFrame implements ActionListener {
 	private JFileChooser fileChooser;
 	private BancoDeDados banco;
 
-	public InputListenerListarEquipamentoFrame(ListarEquipamentoFrame listarEquipamentoFrame,BancoDeDados banco) {
+	public InputListenerListarEquipamentoFrame(ListarEquipamentoFrame listarEquipamentoFrame, BancoDeDados banco) {
 		this.listarEquipamento = listarEquipamentoFrame;
 		this.banco = banco;
 	}
+
 	@Override
 	public void actionPerformed(ActionEvent e) {
 		if (e.getActionCommand().equals("Voltar")) {
@@ -33,24 +34,44 @@ public class InputListenerListarEquipamentoFrame implements ActionListener {
 			new MenuPesquisadorFrame().setVisible(true);
 		} else if (e.getActionCommand().equals("Remover")) {
 			if (listarEquipamento.getList().getSelectedIndex() != -1) {
-				String[] campos = listarEquipamento.getList().getSelectedValue().split(" ");
-				try {
-					System.out.println("campos[2] = " + campos[0]);
-					EquipamentoDAO.getInstance().deleteEquipamento(campos[0]);
-					listarEquipamento.refreshList();
-					JOptionPane.showMessageDialog(null, "Deletado!");
-					if(listarEquipamento.getList().getSelectedIndex() == -1){
-						listarEquipamento.dispose();
-						new MenuPesquisadorFrame().setVisible(true);
+				int i = JOptionPane.showConfirmDialog(null, "Tem certeza que deseja remover?", "Saída",
+						JOptionPane.YES_NO_OPTION);
+				if (i == JOptionPane.YES_OPTION) {
+					if (listarEquipamento.getList().getSelectedIndex() != -1) {
+						String[] campos = listarEquipamento.getList().getSelectedValue().split(" -- ");
+						try {
+							EquipamentoDAO.getInstance().deleteEquipamento(campos[0]);
+							listarEquipamento.refreshList();
+							JOptionPane.showMessageDialog(null, "Deletado!");
+							if (listarEquipamento.getList().getSelectedIndex() == -1) {
+								listarEquipamento.dispose();
+								new MenuPesquisadorFrame().setVisible(true);
+							}
+						} catch (SQLException r) {
+							JOptionPane.showMessageDialog(null, r.getMessage());
+						}
 					}
-				} catch (SQLException r) {
-					JOptionPane.showMessageDialog(null, r.getMessage());
-				}
+				} else
+					JOptionPane.showMessageDialog(null, "Nenhum elemento selecionado.");
 			}
 		} else if (e.getActionCommand().equals("Editar")) {
 			if (listarEquipamento.getList().getSelectedIndex() != -1) {
-				String[] campos = listarEquipamento.getList().getSelectedValue().split(" ");
+				String[] campos = listarEquipamento.getList().getSelectedValue().split(" -- ");
+
+				int aux = 0;
+
+				System.out.println("campos[0]" + campos[0]);
+				System.out.println("campos[1]" + campos[1]);
+				System.out.println("campos[2]" + campos[2]);
+				System.err.println("oioioio");
+				System.err.println(campos.length);
+				while (aux != campos.length) {
+					System.err.println(campos[aux]);
+					aux++;
+				}
+
 				Equipamento equip = EquipamentoDAO.getInstance().getEquipamento(campos[0]);
+				System.out.println(equip.getNome());
 				listarEquipamento.dispose();
 				new EditarEquipamentoFrame(equip).setVisible(true);
 			} else
@@ -59,11 +80,11 @@ public class InputListenerListarEquipamentoFrame implements ActionListener {
 			fileChooser = new JFileChooser();
 			fileChooser.setDialogTitle("Escolha um local para salvar");
 			int retorno = fileChooser.showSaveDialog(null);
-			if(retorno == JFileChooser.APPROVE_OPTION){
+			if (retorno == JFileChooser.APPROVE_OPTION) {
 				String path = String.valueOf(fileChooser.getSelectedFile());
 				GerarPDF gerarPDF = new GerarPDF(banco, path);
 				try {
-					gerarPDF.gerarRelatorioPoder();
+					gerarPDF.gerarRelatorioEquipamento();
 				} catch (IOException | DocumentException e1) {
 					e1.printStackTrace();
 				}
