+++ ProjectICTLab/src/Controller/InputListenerListarEventoFrame.java
@@ -22,10 +22,11 @@ public class InputListenerListarEventoFrame implements ActionListener {
 	private JFileChooser fileChooser;
 	private BancoDeDados banco;
 
-	public InputListenerListarEventoFrame(ListarEventosFrame listarEvento,BancoDeDados banco) {
+	public InputListenerListarEventoFrame(ListarEventosFrame listarEvento, BancoDeDados banco) {
 		this.listarEvento = listarEvento;
 		this.banco = banco;
 	}
+
 	@Override
 	public void actionPerformed(ActionEvent e) {
 		if (e.getActionCommand().equals("Voltar")) {
@@ -33,24 +34,35 @@ public class InputListenerListarEventoFrame implements ActionListener {
 			new MenuPesquisadorFrame().setVisible(true);
 		} else if (e.getActionCommand().equals("Remover")) {
 			if (listarEvento.getList().getSelectedIndex() != -1) {
-				String[] campos = listarEvento.getList().getSelectedValue().split(" ");
-				try {
-					EventoDAO.getInstance().deleteEvento(campos[0]);
-					listarEvento.refreshList();
-					JOptionPane.showMessageDialog(null, "Deletado!");
-					if(listarEvento.getList().getSelectedIndex() == -1){
-						listarEvento.dispose();
-						new MenuPesquisadorFrame().setVisible(true);
+				int i = JOptionPane.showConfirmDialog(null, "Tem certeza que deseja remover?", "Saída",
+						JOptionPane.YES_NO_OPTION);
+				if (i == JOptionPane.YES_OPTION) {
+
+					String[] campos = listarEvento.getList().getSelectedValue().split(" -- ");
+					try {
+						EventoDAO.getInstance().deleteEvento(campos[0]);
+						listarEvento.refreshList();
+						JOptionPane.showMessageDialog(null, "Deletado!");
+						if (listarEvento.getList().getSelectedIndex() == -1) {
+							listarEvento.dispose();
+							new MenuPesquisadorFrame().setVisible(true);
+						}
+					} catch (SQLException r) {
+						JOptionPane.showMessageDialog(null, r.getMessage());
 					}
-				} catch (SQLException r) {
-					JOptionPane.showMessageDialog(null, r.getMessage());
-				}
+				} else
+					JOptionPane.showMessageDialog(null, "Nenhum elemento selecionado.");
 			}
 		} else if (e.getActionCommand().equals("Editar")) {
 			if (listarEvento.getList().getSelectedIndex() != -1) {
-				String[] campos = listarEvento.getList().getSelectedValue().split(" ");
+				String[] campos = listarEvento.getList().getSelectedValue().split(" -- ");
 				Evento event = EventoDAO.getInstance().getEvento(campos[0]);
 				listarEvento.dispose();
+
+				System.out.println("campos[0]" + campos[0]);
+				System.out.println("campos[1]" + campos[1]);
+				System.out.println("campos[2]" + campos[2]);
+
 				new EditarEventoFrame(event).setVisible(true);
 			} else
 				JOptionPane.showMessageDialog(null, "Nenhuma linha selecionada!");
@@ -58,11 +70,11 @@ public class InputListenerListarEventoFrame implements ActionListener {
 			fileChooser = new JFileChooser();
 			fileChooser.setDialogTitle("Escolha um local para salvar");
 			int retorno = fileChooser.showSaveDialog(null);
-			if(retorno == JFileChooser.APPROVE_OPTION){
+			if (retorno == JFileChooser.APPROVE_OPTION) {
 				String path = String.valueOf(fileChooser.getSelectedFile());
 				GerarPDF gerarPDF = new GerarPDF(banco, path);
 				try {
-					gerarPDF.gerarRelatorioPoder();
+					gerarPDF.gerarRelatorioEvento();
 				} catch (IOException | DocumentException e1) {
 					e1.printStackTrace();
 				}
